Task#21 (#26)
* add task - drag and drop task to another column

* status add/delete/move/edit

* .

* Todos and extras

* add todos

* Todos fix

* przypisywanie studenta do zadania

* fix for email and password change

---------

+++ project-rest-api/src/main/java/com/project/controller/projekt/zadanie/ZadanieRestController.java
@@ -1,15 +1,19 @@
 package com.project.controller.projekt.zadanie;
 
+import com.project.auth.AuthenticationService;
 import com.project.controller.projekt.zadanie.requests.ChangeZadanieStatusRequest;
 import com.project.controller.projekt.zadanie.requests.SetZadanieRequest;
 import com.project.model.*;
 import com.project.service.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.Console;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -18,7 +22,7 @@ import java.util.Set;
 @CrossOrigin
 @RequestMapping("/api/v1/projekty/")
 public class ZadanieRestController {
-
+    private static final Logger logger = LoggerFactory.getLogger(AuthenticationService.class);
     private final ProjektService projektService;
     private final ZadanieService zadanieService;
     private final StatusService statusService;
@@ -62,6 +66,7 @@ public class ZadanieRestController {
                 .nazwa(request.getNazwa())
                 .opis(request.getOpis())
                 .status(targetStatus.get())
+                .piorytet(request.getPiorytet())
                 .projekt(projekt)
                 .build();
         Zadanie createdZadanie = zadanieService.setZadanie(zadanie);
@@ -79,17 +84,23 @@ public class ZadanieRestController {
         if (targetStatus.isEmpty()) {
             return ResponseEntity.badRequest().body("Status not found");
         }
-        Set<Student> students = studentService.getStudentsByIds(request.getStudentIds()); // Fetch students from database
+        Set<Student> students = null;
+//        logger.info("studenci"+request.getStudentIds().toString());
+        if (request.getStudentIds() != null) {
+            students = studentService.getStudentsByIds(request.getStudentIds()); // Fetch students from database only if studentIds is not null
+        }
         zadanie.get().setNazwa(request.getNazwa());
         zadanie.get().setOpis(request.getOpis());
         zadanie.get().setStatus(targetStatus.get());
         zadanie.get().setPiorytet(request.getPiorytet());
-        zadanie.get().setStudenci(students);
+        if (students != null) {
+            zadanie.get().setStudenci(students);
+        }
         Zadanie updatedZadanie = zadanieService.setZadanie(zadanie.get());
         return ResponseEntity.ok(updatedZadanie);
     }
 
-    @PreAuthorize("hasAnyRole('NAUCZYCIEL', 'ADMIN')")
+    @PreAuthorize("hasAnyRole('NAUCZYCIEL', 'ADMIN', 'STUDENT')")
     @PatchMapping("/task/status")
     public ResponseEntity<?> updateZadanieStatus(@RequestBody ChangeZadanieStatusRequest request,
                                                  @AuthenticationPrincipal User currentUser) {
@@ -98,12 +109,17 @@ public class ZadanieRestController {
             return ResponseEntity.badRequest().body("Zadanie not found");
         }
         Projekt projekt = zadanie.get().getProjekt();
-        if (!currentUser.getStudent().getProjekty().stream()
-                .anyMatch(proj -> proj.getProjektId().equals(projekt.getProjektId()))) {
-            return ResponseEntity.badRequest().body("Student is not a member of the project");
-        } else if (!currentUser.getTeacher().getProjekty().stream()
-                .anyMatch(proj -> proj.getProjektId().equals(projekt.getProjektId()))) {
-            return ResponseEntity.badRequest().body("Teacher is not a owner of the project");
+        if (currentUser.getRole().equals(Role.NAUCZYCIEL)) {
+            //TODO utworzyłam nauczycielem projekt, a potem nie mogłam zmieniąc ststusów zadań
+//            if (projekt.getTeacher() != currentUser.getTeacher()) {
+//                return ResponseEntity.badRequest().body("You are not allowed to change status of this task");
+//            }
+        }
+        if (currentUser.getRole().equals(Role.STUDENT)) {
+            //TODO utworzyłam nauczycielem projekt, a potem nie mogłam zmieniąc ststusów zadań
+//            if (projekt.getTeacher() != currentUser.getTeacher()) {
+//                return ResponseEntity.badRequest().body("You are not allowed to change status of this task");
+//            }
         }
         Optional<Status> targetStatus = statusService.getStatus(request.getStatusId());
         if (targetStatus.isEmpty()) {
@@ -126,6 +142,7 @@ public class ZadanieRestController {
         return ResponseEntity.ok().build();
     }
 
+    //TODO odpisywanie się nie działa. Dostaję 200, ale nic się nie zmienia
     @PreAuthorize("hasRole('STUDENT')")
     @PostMapping("/task/{zadanieId}/leave")
     public ResponseEntity<?> leaveZadanie(@PathVariable Long zadanieId, @AuthenticationPrincipal User currentUser) {
@@ -141,6 +158,15 @@ public class ZadanieRestController {
     @PreAuthorize("hasAnyRole('NAUCZYCIEL', 'ADMIN')")
     @DeleteMapping("/task/{zadanieId}")
     public ResponseEntity<?> deleteZadanie(@PathVariable Long zadanieId) {
+        Optional<Zadanie> zadanie = zadanieService.getZadanie(zadanieId);
+        if (zadanie.isEmpty()) {
+            return ResponseEntity.badRequest().body("Zadanie not found");
+        }
+
+        // Unassign all students from the zadanie
+        zadanie.get().getStudenci().clear();
+        zadanieService.setZadanie(zadanie.get());
+
         zadanieService.deleteZadanie(zadanieId);
         return ResponseEntity.ok().build();
     }
